% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adding.R
\name{adding}
\alias{adding}
\title{Adding two integers}
\usage{
adding(x, y)
}
\arguments{
\item{x}{An integer}

\item{y}{An integer}
}
\value{
An integer.
}
\description{
Adding two integers
}
\examples{
x <- 1
y <- 2
adding(1,2)
}
