% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{\strong{fars_map_state}}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{number representing a US state as coded by the US National Highway Traffic
Safety Administration'sFatality Analysis Reporting System}

\item{year}{vector of 4-digit numeric numbers for a specied data years from US National Highway Traffic
Safety Administration's Fatality Analysis Reporting System}
}
\value{
if data is available for the queried state and year, opens a  plot window showing the state outline
and location of reported accidents across the year.  If no data, will return warning.
}
\description{
Function to plot fatality locations for a specified year and state. \emph{Requires map from maps, points from graphics,}
\emph{and filter from dply packages}.
}
\examples{
\dontrun{
fars_map_state(16, 2013) # where 16=Idaho and 2013 is the queried year
}


}
