% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{\strong{fars_read}}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{full name of the file containing data to be imported into R session}
}
\value{
Returns a data.frame object of the class tbl_df (also known as a \emph{tibble}--
the structure central data of  tidyverse packages.  If the file is not located, will return an error
stating that filename does not exist.
}
\description{
This function reads in data using the readr package. \emph{Requires the readr package}
or the function will fail.
}
\note{
a vector of multiple filenames can be queried.  Output combines files into one tibble.
}
\examples{
\dontrun{
data.13 <- fars_read("accident_2013.csv.bz2")
creates a tibble from the file  "accident_2013.csv.bz2"
}

}
