% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{\strong{fars_read_years}}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{vector of 4-digit numeric numbers for a specied data years from US National Highway Traffic
Safety Administration'sFatality Analysis Reporting System}
}
\value{
if the file exists, creates a list consisting of a 2-column tibble for each queried year, with month (1-12)
and 4-digit year for each row data in the queried file.  If the file is not located, will include a NULL value
for that year
}
\description{
Function to read compressed FARS csv files for a range of years and provides a month (1-12) and 4-digit year for each line item
included in each file.  \emph{Requires mutate and select from the dplyr package}
}
\examples{
\dontrun{
fars_read_years(2013:2015)
}

}
