% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{\strong{fars_summarize_years}}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a numeric vector of 4-digit numbers for specified data years from US National Highway Traffic
Safety Administration'sFatality Analysis Reporting System}
}
\value{
Creates a tibble summarizing the total number of accidents reported by month and year
}
\description{
Function to summarize results from \emph{fars_read_years} function.  \emph{Requires group_by and summarize from dplyr and}
\emph{Requires group_by and summarize from dplyr and spread from tidyr packages.}
a
}
\examples{
\dontrun{
fars_summarize_years(2013:2015)
}

}
