% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{\strong{make_filename}}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a 4-digit numeric number for a specied data year from US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System}
}
\value{
outputs a character value of "accident_\emph{year}.csv.bz2"
}
\description{
Simple function to create a filename of a compressed csv FARS file for a specified year, with a .bz2 extension.
}
\examples{
\dontrun{
make.filename(2013)  # outputs "accident_2013.csv.bz2"
}

}
